Rule out words with the wrong length.
When looking for anagrams of target_word, we can rule out any words that
do not have the same length as target_word. In doing so, we can save
ourselves the work of checking if one word is an anagram of another.

Doing this is worthwhile only if checking the word length is a faster
operation than checking if one word is an anagram of another. With any
resonable implementation of both operations, this must be the case, and
thus ruling out words with "the wrong lenght" must save us time in
practice. Indeed, in my tests, with an input file with approx. 400k
words, the running time was cut down by half.